Moved utility tests and added std::swap tests
