+++ kernel/UnitTests/KernelStdlib/UtilityTests.h
@@ -1,0 +1,12 @@
+#ifndef KERNEL_UNITTESTS_KERNELSTDLIB_UTILITYTESTS_H
+#define KERNEL_UNITTESTS_KERNELSTDLIB_UTILITYTESTS_H
+
+namespace UnitTests::KernelStdlib::Utility
+{
+    /**
+     * Run all runtime tests
+     */
+    void Run();
+}
+
+#endif // KERNEL_UNITTESTS_KERNELSTDLIB_UTILITYTESTS_H
